move initial state into module
